+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!--next-version-placeholder-->
 
+## v0.1.0 (26/05/2025)
+
+- Modularização do TJSP
+- Tribunais adicionados: TJRS, TJPR, TJDFT (apenas CJSG)
+- Agregadores adicionados: Datajud, Jusbr
+
 ## v0.0.1 (17/12/2024)
 
 - First release of `juscraper`!